requirements.txt
PyPDF2
numpy
faiss-cpu
sentence-transformers
python-dotenv
PyCryptodome
google-generativeai
ipykernel
streamlit